fix: 수익률 계산을 BigDecimal로 변경하여 정확성 개선

+++ src/main/java/lotto/utils/LottoGenerator.java
@@ -1,6 +1,7 @@
 package lotto.utils;
 
 import camp.nextstep.edu.missionutils.Randoms;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.stream.Stream;
 import lotto.domain.Lotto;
@@ -9,9 +10,9 @@ import lotto.domain.Number;
 
 public class LottoGenerator {
 
-    public static LottoTickets generateLottoTickets(int lottoAmount) {
+    public static LottoTickets generateLottoTickets(BigDecimal lottoAmount) {
         List<Lotto> lottoList = Stream.generate(LottoGenerator::generateSortedLottoNumbers)
-                .limit(lottoAmount)
+                .limit(lottoAmount.longValue())
                 .map(Lotto::new)
                 .toList();
